Remove seperate props classes
