--- components/cool-text/cool-text.props.ts
@@ -1,9 +1,0 @@
-import { TypographyVariant } from "@mui/material";
-
-export default interface CoolTextProps {
-  text: string;
-  inline?: boolean;
-  caption?: string;
-  forcedHoverState?: boolean;
-  variant: TypographyVariant;
-}
