--- components/default-layout-wrapper/default-layout-wrapper.props.ts
@@ -1,8 +1,0 @@
-import { Fab } from "@mui/material";
-import { ReactNode } from "react";
-
-export default interface DefaultLayoutWrapperProps {
-  children: ReactNode;
-  fab?: typeof Fab;
-  removeBackgroundBlur?: boolean;
-}
