--- components/home/personal-projects-section/personal-projects-section.props.ts
@@ -1,5 +1,0 @@
-import { PersonalProject } from "@/interfaces";
-
-export default interface PersonalProjectsSectionProps {
-  projects: PersonalProject[];
-}
