--- components/home/work-experience-section/work-experience-section.props.ts
@@ -1,5 +1,0 @@
-import { WorkExperience } from "@/interfaces";
-
-export default interface WorkExperienceSectionProps {
-  workExperiences: WorkExperience[];
-}
