--- components/project-link-button/project-link-button.props.ts
@@ -1,6 +1,0 @@
-import { ProjectLink } from "@/interfaces";
-
-export default interface ProjectLinkButtonProps {
-  name: string;
-  link: ProjectLink;
-}
