--- components/timeline/mobile/timeline/mobile-timeline.props.ts
@@ -1,5 +1,0 @@
-import { WorkExperience } from "@/interfaces";
-
-export default interface MobileTimelineProps {
-  experiences: WorkExperience[];
-}
