--- components/timeline/work-experience-timeline-item/work-experience-timeline-item.props.ts
@@ -1,6 +1,0 @@
-import { WorkExperience } from "@/interfaces";
-
-export default interface WorkExperienceTimelineItemProps
-  extends WorkExperience {
-  index: number;
-}
