--- components/timeline/work-experience-timeline/work-experience-timeline.props.ts
@@ -1,5 +1,0 @@
-import { WorkExperience } from "@/interfaces";
-
-export default interface WorkExperienceTimelineProps {
-  experiences: WorkExperience[];
-}
