--- components/work-experience-card/work-experience-card.props.ts
@@ -1,15 +1,0 @@
-import { WorkExperience } from "@/interfaces";
-
-export default interface WorkExperienceCardProps extends WorkExperience {
-  minHeight?: boolean;
-  margin?: boolean;
-  maxWidth?: boolean;
-  justifyContent:
-    | "center"
-    | "end"
-    | "flex-end"
-    | "flex-start"
-    | "start"
-    | undefined;
-  shadow?: boolean;
-}
