--- components/work-project-card/work-project-card.props.ts
@@ -1,7 +1,0 @@
-import { WorkProject } from "@/interfaces/workProject.interface";
-
-export default interface WorkProjectCardProps {
-  project: WorkProject;
-  maxDescriptionHeight: number;
-  maxTechSectionHeight: number;
-}
